Use the PR titles when creating the automatic update pull-requests.
The update.py script would generate a changelog that included all of the
commits titles. Since we don't have a culture of squashing pull requests
before merging, the list of commits is rather long, and most titles are
meaningless.

We get a much better signal-to-noise ratio by using the list of pull
request titles instead. We cannot get that list directly, but we can map
each commit to its corresponding PR and use that instead. The pull
requests are only listed once, even if referenced by multiple commits,
leading to a much shorter changelog.