Use the vimc-robot account to create the pull request.
Pull request created by an action's default GITHUB_TOKEN do not trigger
any further action runs (presumably to avoid cycles). This was affecting
the nightly update action, and required manual intervention to get the
CI to run.

By using a token that is tied to a specific account we can workaround
this. I have created a fork of this repository under
vimc-robot/packit-infra and added a PAT that can push to it and create
pull requests.